A simple Rust-C FFI app
Added a simple Rust-C FFI app to read chunks from a source text file.
`bindgen` was used to create Rust bindings for a C library, and a Rust
module was written to expose relevant parts of the library. A main
application was written to demonstrate and exercise the API.
I wrote this in order to learn more about how FFI works in Rust.